requirements.txt
openai
pytesseract
pillow